Split controller and settings
