+++ apps/updatenotification/lib/Controller/AdminController.php
@@ -26,21 +26,17 @@
 namespace OCA\UpdateNotification\Controller;
 
 use OCA\UpdateNotification\ResetTokenBackgroundJob;
-use OCA\UpdateNotification\UpdateChecker;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http\DataResponse;
-use OCP\AppFramework\Http\TemplateResponse;
 use OCP\AppFramework\Utility\ITimeFactory;
 use OCP\BackgroundJob\IJobList;
 use OCP\IConfig;
-use OCP\IDateTimeFormatter;
 use OCP\IL10N;
 use OCP\IRequest;
 use OCP\Security\ISecureRandom;
-use OCP\Settings\ISettings;
 use OCP\Util;
 
-class AdminController extends Controller implements ISettings {
+class AdminController extends Controller {
 	/** @var IJobList */
 	private $jobList;
 	/** @var ISecureRandom */
@@ -49,12 +45,8 @@ class AdminController extends Controller implements ISettings {
 	private $config;
 	/** @var ITimeFactory */
 	private $timeFactory;
-	/** @var UpdateChecker */
-	private $updateChecker;
 	/** @var IL10N */
 	private $l10n;
-	/** @var IDateTimeFormatter */
-	private $dateTimeFormatter;
 
 	/**
 	 * @param string $appName
@@ -64,8 +56,6 @@ class AdminController extends Controller implements ISettings {
 	 * @param IConfig $config
 	 * @param ITimeFactory $timeFactory
 	 * @param IL10N $l10n
-	 * @param UpdateChecker $updateChecker
-	 * @param IDateTimeFormatter $dateTimeFormatter
 	 */
 	public function __construct($appName,
 								IRequest $request,
@@ -73,60 +63,13 @@ class AdminController extends Controller implements ISettings {
 								ISecureRandom $secureRandom,
 								IConfig $config,
 								ITimeFactory $timeFactory,
-								IL10N $l10n,
-								UpdateChecker $updateChecker,
-								IDateTimeFormatter $dateTimeFormatter) {
+								IL10N $l10n) {
 		parent::__construct($appName, $request);
 		$this->jobList = $jobList;
 		$this->secureRandom = $secureRandom;
 		$this->config = $config;
 		$this->timeFactory = $timeFactory;
 		$this->l10n = $l10n;
-		$this->updateChecker = $updateChecker;
-		$this->dateTimeFormatter = $dateTimeFormatter;
-	}
-
-	/**
-	 * @return TemplateResponse
-	 */
-	public function getForm() {
-		$lastUpdateCheckTimestamp = $this->config->getAppValue('core', 'lastupdatedat');
-		$lastUpdateCheck = $this->dateTimeFormatter->formatDateTime($lastUpdateCheckTimestamp);
-
-		$channels = [
-			'daily',
-			'beta',
-			'stable',
-			'production',
-		];
-		$currentChannel = Util::getChannel();
-
-		// Remove the currently used channel from the channels list
-		if(($key = array_search($currentChannel, $channels)) !== false) {
-			unset($channels[$key]);
-		}
-		$updateState = $this->updateChecker->getUpdateState();
-
-		$notifyGroups = json_decode($this->config->getAppValue('updatenotification', 'notify_groups', '["admin"]'), true);
-
-		$defaultUpdateServerURL = 'https://updates.nextcloud.com/server/';
-		$updateServerURL = $this->config->getSystemValue('updater.server.url', $defaultUpdateServerURL);
-
-		$params = [
-			'isNewVersionAvailable' => !empty($updateState['updateAvailable']),
-			'isUpdateChecked' => $lastUpdateCheckTimestamp > 0,
-			'lastChecked' => $lastUpdateCheck,
-			'currentChannel' => $currentChannel,
-			'channels' => $channels,
-			'newVersionString' => (empty($updateState['updateVersion'])) ? '' : $updateState['updateVersion'],
-			'downloadLink' => (empty($updateState['downloadLink'])) ? '' : $updateState['downloadLink'],
-			'updaterEnabled' => (empty($updateState['updaterEnabled'])) ? false : $updateState['updaterEnabled'],
-			'isDefaultUpdateServerURL' => $updateServerURL === $defaultUpdateServerURL,
-			'updateServerURL' => $updateServerURL,
-			'notify_groups' => implode('|', $notifyGroups),
-		];
-
-		return new TemplateResponse($this->appName, 'admin', $params, '');
 	}
 
 	/**
@@ -152,23 +95,5 @@ class AdminController extends Controller implements ISettings {
 		$this->config->setSystemValue('updater.secret', password_hash($newToken, PASSWORD_DEFAULT));
 
 		return new DataResponse($newToken);
-	}
-
-	/**
-	 * @return string the section ID, e.g. 'sharing'
-	 */
-	public function getSection() {
-		return 'server';
-	}
-
-	/**
-	 * @return int whether the form should be rather on the top or bottom of
-	 * the admin section. The forms are arranged in ascending order of the
-	 * priority values. It is required to return a value between 0 and 100.
-	 *
-	 * E.g.: 70
-	 */
-	public function getPriority() {
-		return 1;
 	}
 }
